requirements.txt
# ==================== ОСНОВНЫЕ ЗАВИСИМОСТИ ====================

# Telegram Bot Framework
aiogram==3.15.0

# Async PostgreSQL Driver (для продакшн развёртывания)
asyncpg==0.30.0

# Async SQLite Driver (для локальной разработки и тестирования)
aiosqlite==0.20.0

# OpenAI API (совместим с OpenRouter для DeepSeek R1)
openai==1.59.5

# Environment Variables
python-dotenv==1.0.1

# ==================== MACHINE LEARNING (RuBERT) ====================

# Hugging Face Transformers (для RuBERT классификатора)
transformers==4.36.0

# PyTorch (для inference RuBERT модели на CPU)
torch==2.1.0

# NLP обработка текста (keyword extraction для СЛОЯ 2)
nltk==3.8

# Морфологический анализ для русского языка
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417150.4580142

# ==================== ВСПОМОГАТЕЛЬНЫЕ БИБЛИОТЕКИ ====================

# Асинхронный HTTP клиент (для DeepSeek API в СЛОЕ 3)
httpx==0.28.1

# Асинхронная работа с файлами
aiofiles==24.1.0

# Парсинг дат
python-dateutil==2.9.0

# JSON валидация и конфигурация
pydantic==2.9.2

# ==================== ОПЦИОНАЛЬНО ====================

# Мониторинг системы (CPU, RAM, uptime для команды /stats)
psutil==6.1.0

# ==================== ПРИМЕЧАНИЯ ====================
# 
# УСТАНОВКА:
# pip install -r requirements.txt
#
# МИНИМАЛЬНАЯ УСТАНОВКА (без RuBERT, только DeepSeek API fallback):
# pip install aiogram aiosqlite openai python-dotenv httpx
# ВНИМАНИЕ: Без RuBERT все запросы будут обрабатываться через DeepSeek API (медленно, ~2-4 сек)
#
# ПОЛНАЯ УСТАНОВКА (с RuBERT классификатором - рекомендуется):
# pip install -r requirements.txt
# Преимущества: 87% запросов обрабатываются за <100 мс через RuBERT + FAQ
#
# СИСТЕМНЫЕ ТРЕБОВАНИЯ:
# Минимальные (разработка):
# - Python 3.10 или выше
# - pip последней версии: python -m pip install --upgrade pip
# - CPU: 2 ядра, 2.0 GHz
# - RAM: 4 GB
# - HDD: 20 GB SSD
# - Интернет для загрузки зависимостей
#
# Рекомендуемые (продакшн с RuBERT):
# - Python 3.10 или выше
# - CPU: 4 ядра, 3.0 GHz
# - RAM: 16 GB (для RuBERT модели ~700 MB постоянно в памяти)
# - SSD: 100 GB (база данных + RuBERT модель + логи)
# - Интернет: стабильное подключение минимум 10 Мбит/с
#
# VIRTUAL ENVIRONMENT (строго рекомендуется):
# python -m venv venv
# Windows: venv\Scripts\activate
# Linux/Mac: source venv/bin/activate
#
# ПРИМЕЧАНИЕ ПРО PYTORCH:
# Для CPU (по умолчанию): torch==2.1.0 установится автоматически (~800 MB)
# Для GPU с CUDA (опционально): см. https://pytorch.org/get-started/locally/
# Примечание: RuBERT работает достаточно быстро на CPU (50-100 мс), GPU не обязателен
#
# РАЗМЕР УСТАНОВКИ:
# - Базовые зависимости: ~200 MB
# - PyTorch (CPU): ~800 MB
# - Transformers: ~50 MB
# - RuBERT модель: ~700 MB
# - Итого: ~1.8 GB
#
# ПЕРВЫЙ ЗАПУСК:
# После установки зависимостей при первом запуске бота:
# 1. RuBERT модель загрузится в память (~5-10 секунд)
# 2. Первый запрос может занять ~5-10 секунд
# 3. Все последующие запросы: ~50-100 миллисекунд (СЛОЙ 1: RuBERT)
